% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curriculum_graph_from_json.R
\name{curriculum_graph_from_json}
\alias{curriculum_graph_from_json}
\title{Create Curriculum From JSON File}
\usage{
curriculum_graph_from_json(file)
}
\arguments{
\item{file}{A json file following the format produced by \code{\link[=curriculum_graph_to_json]{curriculum_graph_to_json()}}}
}
\value{
A list that contains the following: \item{node_list}{A dataframe of
course nodes containg at least their id, term, blocking factor (bf), delay
factor (df), centrality (cf), and cruciality (sc)} \item{edge_list}{A
dataframe with two columns 'from' and 'to' specifying directed edges
starting at 'from' nodes directed towards 'to' nodes.} \item{network}{Igraph
network object representing the curriculum graph} \item{sc_total}{Total
structural complexity of the curriculum graph} \item{bf_total}{Total
blocking factor of the curriculum graph} \item{df_total}{Total delay factor
of the curriculum graph}
}
\description{
Generates a curriculum graph from a json file.
}
