% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curriculum_graph_to_json.R
\name{curriculum_graph_to_json}
\alias{curriculum_graph_to_json}
\title{Write curriculum graph object to JSON}
\usage{
curriculum_graph_to_json(curriculum_graph, file_path)
}
\arguments{
\item{curriculum_graph}{A curriculum_graph object created with either
\code{\link[=curriculum_graph_from_list]{curriculum_graph_from_list()}} or \code{\link[=curriculum_graph_from_json]{curriculum_graph_from_json()}}}

\item{file_path}{A path to the JSON file you want written. If the file does
not exist in the specified directory one will be created.}
}
\description{
A function to save a curriculum graph object in a JSON file format.
}
\examples{
edge_list <- data.frame(from = c(1, 2, 3), to = c(4, 4, 4))
# courses in node list must be placed sequentially in term order to be properly displayed
node_list <-
data.frame(
 id = 1:4,
 label = c("MATH 100", "DATA 101", "MATH 101", "MATH 221"),
 term = c(1, 1, 2, 2)
)
file_path <- "./test.json"
C <- curriculum_graph_from_list(node_list, edge_list)
#curriculum_graph_to_json(C, "./test.json")
}
\references{
Hickman, Michael S. 2017."Development of a curriculum analysis and
simulation library with applications in curricular analytics."
}
\author{
Daniel Krasnov
}
