% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcads.R
\name{dcads}
\alias{dcads}
\title{Launch DCADS: Data-driven Curriculum Analytics and Design System}
\usage{
dcads(
  curriculum_csv,
  topic_model = NULL,
  pred_model = NULL,
  front_text = NULL,
  course_codes = NULL
)
}
\arguments{
\item{curriculum_csv}{A CSV in the form of \code{filepath} found in \link{curriculum_graph_from_csv}}
}
\value{
...
}
\description{
A function to create and launch the DCADS Shiny app.
}
\details{
More explanation here...
}
\examples{
...

}
\references{
Heileman, Gregory L, Chaouki T Abdallah, Ahmad Slim, and Michael
Hickman. 2018. “Curricular Analytics: A Framework for Quantifying the Impact
of Curricular Reforms and Pedagogical Innovations.” arXiv Preprint
arXiv:1811.09676.
}
\author{
Daniel Krasnov
}
